requirements.txt
argparse==1.2.1
beautifulsoup==3.2.1
cherrypy==3.5.0
cov-core==1.14.0
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-colorful==1.0.1
django-debug-toolbar==1.2.1
django-extensions==1.4.0
django-pipeline==1.3.25
django-pylibmc-sasl==0.2.4
django-taggit==0.12.1
django==1.7
fabric==1.10.0
factory-boy==2.4.1
flask==0.10.1
fuzzywuzzy==0.3.1
git+https://github.com/nvie/pip-tools.git@e17361ecf18bf459e20c9c1410a7447232db7600#egg=pip-tools
glob2==0.4.1
gunicorn==19.1.1
icalendar==3.8.3
itsdangerous==0.24
jasmine-core==2.0.2
jasmine==2.0.0
jinja2==2.7.3
lettuce==0.2.19
markupsafe==0.23
mock==1.0.1
mockfs==1.0.1
ordereddict==1.1
paramiko==1.14.1
pep8==1.5.7
psycopg2==2.5.4
py==1.4.24
pyconstraints==1.0.1
pylibmc==1.3.0
pypdf2==1.23
pytest-cov==1.8.0
pytest-django==2.6.2
pytest==2.6.2
python-dateutil==2.2
pytz==2014.7
pyyaml==3.10
rpicourses==1.0.8
selenium==2.42.1
setuptools==5.7
six==1.7.3
sqlparse==0.1.11
static3==0.5.1
sure==1.2.7
werkzeug==0.9.6
